requirements.txt
langchain>=0.2.0
langchain-community>=0.2.0
langchain-mistralai>=0.1.0
chromadb>=0.5.0
python-dotenv>=1.0.0
